feat: Move DoorRepositoryImpl and PushRepositoryImpl to data module (Phase 9)
Move repository implementations from androidApp to data/src/commonMain/:
- DoorRepositoryImpl → data/repository/ (zero Android deps)
- PushRepositoryImpl → data/repository/ (JVM deps: Duration, Date —
  works for Android KMP target, needs cleanup for iOS)
- createButtonAckToken() stays in androidApp (uses SimpleDateFormat)
- Add Kermit dependency to data module

All 3 repository implementations now live in the shared data module:
ServerConfigRepositoryImpl, DoorRepositoryImpl, PushRepositoryImpl.
