+++ AndroidGarage/data/build.gradle.kts
@@ -10,6 +10,7 @@ kotlin {
         commonMain.dependencies {
             implementation(project(":domain"))
             implementation(libs.kotlinx.coroutines.core)
+            implementation(libs.kermit)
         }
     }
 }
