+++ AndroidGarage/data/src/commonMain/kotlin/com/chriscartland/garage/data/repository/DoorRepositoryImpl.kt
@@ -15,7 +15,7 @@
  *
  */
 
-package com.chriscartland.garage.door
+package com.chriscartland.garage.data.repository
 
 import co.touchlab.kermit.Logger
 import com.chriscartland.garage.data.LocalDoorDataSource
