+++ AndroidGarage/data/src/commonMain/kotlin/com/chriscartland/garage/data/repository/PushRepositoryImpl.kt
@@ -15,7 +15,7 @@
  *
  */
 
-package com.chriscartland.garage.remotebutton
+package com.chriscartland.garage.data.repository
 
 import co.touchlab.kermit.Logger
 import com.chriscartland.garage.data.NetworkButtonDataSource
@@ -27,7 +27,6 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.time.delay
 import java.time.Duration
-import java.util.Date
 
 class PushRepositoryImpl(
     private val networkButtonDataSource: NetworkButtonDataSource,
@@ -119,22 +118,4 @@ class PushRepositoryImpl(
         }
         _snoozeRequestStatus.value = SnoozeRequestStatus.IDLE
     }
-}
-
-/**
- * Create a button ack token.
- *
- * This token is created by the client so the server can acknowledge the remote button push.
- * The client can send the same token to the server multiple times and the server is
- * responsible for only processing the token once.
- * When the server receives a button press, it will respond with the token to the client.
- */
-fun createButtonAckToken(now: Date): String {
-    val humanReadable = java.text.SimpleDateFormat("yyyy-MM-dd hh:mm:ss a", java.util.Locale.US).format(now)
-    val timestampMillis = now.time
-    val appVersion = "AppVersionTODO"
-    val buttonAckTokenData = "android-$appVersion-$humanReadable-$timestampMillis"
-    val re = Regex("[^a-zA-Z0-9-_.]")
-    val filtered = re.replace(buttonAckTokenData, ".")
-    return filtered
 }
